remove exp
