+++ src/istio-watcher/cmd/main.go
@@ -1,8 +1,8 @@
 package main
 
 import (
-	"github.com/otterize/network-mapper/src/exp/istio-watcher/mapperclient"
-	"github.com/otterize/network-mapper/src/exp/istio-watcher/pkg/watcher"
+	"github.com/otterize/network-mapper/src/istio-watcher/mapperclient"
+	"github.com/otterize/network-mapper/src/istio-watcher/pkg/watcher"
 	sharedconfig "github.com/otterize/network-mapper/src/shared/config"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
@@ -15,7 +15,7 @@ func main() {
 	}
 
 	mapperClient := mapperclient.NewMapperClient(viper.GetString(sharedconfig.MapperApiUrlKey))
-	istioWatcher, err := istiowatcher.NewWatcher(mapperClient)
+	istioWatcher, err := istiowatcher.istiowatcher.NewWatcher(mapperClient)
 	if err != nil {
 		logrus.WithError(err).Panic()
 	}
