+++ src/istio-watcher/mapperclient/generate.go
