+++ src/istio-watcher/mapperclient/generated.go
