+++ src/istio-watcher/mapperclient/genqlient.graphql
