+++ src/istio-watcher/mapperclient/genqlient.yaml
