+++ src/istio-watcher/pkg/watcher/istioconnectionmatcher.go
@@ -2,7 +2,7 @@ package istiowatcher
 
 import (
 	"fmt"
-	"github.com/otterize/network-mapper/src/exp/istio-watcher/mapperclient"
+	"github.com/otterize/network-mapper/src/istio-watcher/mapperclient"
 	"github.com/samber/lo"
 	"golang.org/x/exp/slices"
 )
