+++ src/istio-watcher/pkg/watcher/watcher_test.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 	"github.com/golang/mock/gomock"
-	"github.com/otterize/network-mapper/src/exp/istio-watcher/mapperclient"
-	mock_mapperclient "github.com/otterize/network-mapper/src/exp/istio-watcher/mapperclient/mockclient"
+	"github.com/otterize/network-mapper/src/istio-watcher/mapperclient"
+	"github.com/otterize/network-mapper/src/istio-watcher/mapperclient/mockclient"
 	"github.com/stretchr/testify/suite"
 	"testing"
 	"time"
