+++ src/kafka-watcher/pkg/logwatcher/kubeneteslogwatcher.go
@@ -4,8 +4,8 @@ import (
 	"bufio"
 	"context"
 	"errors"
-	"github.com/otterize/network-mapper/src/exp/kafka-watcher/pkg/config"
-	"github.com/otterize/network-mapper/src/exp/kafka-watcher/pkg/mapperclient"
+	"github.com/otterize/network-mapper/src/kafka-watcher/pkg/config"
+	"github.com/otterize/network-mapper/src/kafka-watcher/pkg/mapperclient"
 	"github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 	"golang.org/x/exp/slices"
