+++ src/kafka-watcher/pkg/logwatcher/watcher.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"errors"
 	"github.com/oriser/regroup"
-	"github.com/otterize/network-mapper/src/exp/kafka-watcher/pkg/mapperclient"
+	mapperclient2 "github.com/otterize/network-mapper/src/kafka-watcher/pkg/mapperclient"
 	"github.com/samber/lo"
 	"github.com/sirupsen/logrus"
 	"k8s.io/apimachinery/pkg/types"
@@ -36,7 +36,7 @@ type Watcher interface {
 type baseWatcher struct {
 	mu           sync.Mutex
 	seen         SeenRecordsStore
-	mapperClient mapperclient.MapperClient
+	mapperClient mapperclient2.MapperClient
 }
 
 func (b *baseWatcher) flush() SeenRecordsStore {
@@ -59,8 +59,8 @@ func (b *baseWatcher) reportResults(ctx context.Context) error {
 
 	logrus.Infof("Reporting %d records", cRecords)
 
-	results := lo.MapToSlice(records, func(r AuthorizerRecord, t time.Time) mapperclient.KafkaMapperResult {
-		return mapperclient.KafkaMapperResult{
+	results := lo.MapToSlice(records, func(r AuthorizerRecord, t time.Time) mapperclient2.KafkaMapperResult {
+		return mapperclient2.KafkaMapperResult{
 			SrcIp:           r.Host,
 			ServerPodName:   r.Server.Name,
 			ServerNamespace: r.Server.Namespace,
@@ -70,7 +70,7 @@ func (b *baseWatcher) reportResults(ctx context.Context) error {
 		}
 	})
 
-	return b.mapperClient.ReportKafkaMapperResults(ctx, mapperclient.KafkaMapperResults{Results: results})
+	return b.mapperClient.ReportKafkaMapperResults(ctx, mapperclient2.KafkaMapperResults{Results: results})
 }
 
 func (b *baseWatcher) processLogRecord(kafkaServer types.NamespacedName, record string) {
