+++ src/kafka-watcher/pkg/mapperclient/genqlient.graphql
