+++ src/kafka-watcher/pkg/mapperclient/genqlient.yaml
